cloud_functions/get_status/requirements.txt
google-cloud-logging==3.5.0
google-cloud-batch==0.5.0
google-cloud-storage==2.9.0
google-cloud-bigquery